Split the kickback and report text out of the verify orchestrator
Third extraction of the same shape: `markup-verify.ts` paired a Playwright
orchestrator with 200 lines of string building over the report it produces.
Importing that module costs 548ms; the string builders need 93ms of it, and
their own tests were already pure and paying the difference to run.

This file needed a different tool than the previous two. Range-slicing
worked for `component-from-image.ts` and `page-compose.ts` because their
pure halves were contiguous; `markup-verify.ts` interleaves imports with
code, so a range grabbed seven import lines along with `kindTag`.
Extracting by declaration name instead — walking back over the doc comment
and forward to the closing brace — is correct here, and it caught its own
bug on the way: a function's parameter list closes with `): string {` in
column 0, so a naive "stop at the first closer" rule truncated a 119-line
function to its 5-line signature. Terminating only on `}` for functions
fixes it, and the line count is what showed it.

The verdict types stay in `markup-verify.ts`. They are that module's public
data model, the same reason `ComponentFromImageReport` stayed put when the
component formatter moved; the type import back is erased, so the runtime
dependency runs one way.
